LayoutTests/imported/w3c/web-platform-tests/FileAPI/reading-data-section/filereader_readAsText-expected.txt


FAIL readAsText should correctly read UTF-8. assert_equals: expected 1 but got 2
PASS readAsText should correctly read UTF-16.
